requirements.txt
Django==5.0.6
django-allauth==0.63.6
django-crispy-forms==2.3
django-environ==0.11.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
